refactor/#37 : 모든 svg 파일 designsystem 모듈에서 관리
