+++ core/designsystem/src/main/res/drawable/ic_eat.xml
