+++ core/designsystem/src/main/res/drawable/ic_group.xml
