+++ core/designsystem/src/main/res/drawable/ic_other.xml
