+++ core/designsystem/src/main/res/drawable/ic_see.xml
