+++ feature/home/src/main/java/com/boostcamp/mapisode/home/common/ChipType.kt
@@ -5,15 +5,16 @@ import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import com.boostcamp.mapisode.designsystem.theme.MapisodeTheme
-import com.boostcamp.mapisode.home.R
+import com.boostcamp.mapisode.home.R as Home
+import com.boostcamp.mapisode.designsystem.R as Design
 
 enum class ChipType(
 	@StringRes val textResId: Int,
 	@DrawableRes val iconResId: Int,
 ) {
-	EATERY(R.string.home_chip_eat, R.drawable.ic_eat),
-	SEEING(R.string.home_chip_see, R.drawable.ic_see),
-	OTHER(R.string.home_chip_other, R.drawable.ic_other),
+	EATERY(Home.string.home_chip_eat, Design.drawable.ic_eat),
+	SEEING(Home.string.home_chip_see, Design.drawable.ic_see),
+	OTHER(Home.string.home_chip_other, Design.drawable.ic_other),
 }
 
 @Composable
